interleaving: clarify types in bit_array_interleave
Follow-up 79f21e017a6c37a60ee1a79d8bea74007f533126

+++ bit_array_interleave.h
@@ -7,7 +7,11 @@
 struct bit_array_interleave_lookup_table {
 	/**
 	 * Contains octet lookup table with
-	 * shifts for each dimension
+	 * shifts for each dimension.
+	 * Structure:
+	 *   - dimensions: [1; 20]
+	 *   - lookup table 2^8 values: [0; 255]
+	 *   - bit array
 	 */
 	bit_array ***tables;
 	/**
@@ -17,7 +21,7 @@ struct bit_array_interleave_lookup_table {
 	/**
 	 * Amount of dimensions
 	 */
-	size_t dim;
+	uint8_t dim;
 	/**
 	 * Mempool ptr
 	 */
@@ -25,7 +29,7 @@ struct bit_array_interleave_lookup_table {
 };
 
 struct bit_array_interleave_lookup_table *
-bit_array_interleave_new_lookup_tables(struct mempool *pool, size_t dim);
+bit_array_interleave_new_lookup_tables(struct mempool *pool, uint8_t dim);
 
 void
 bit_array_interleave_free_lookup_tables(
